Home

SteadyStateFit.jl

SteadyStateFit.SteadyStateObjective

SteadyStateFit.optimize

SteadyStateFit.SteadyStateObjective — Type.

SteadyStateObjective(
    loss,
    ode::ODEProblem,
    conditions,
    conditionsetter::Lens,
    parameterlens::Lens,
)

SteadyStateObjective defines an objective function F

\[F(x) = \frac{1}{N} \sum_{c \in \mathtt{conditions}} \mathtt{loss}(u(x, c), c)\]

where N = length(conditions) and $u(x, c)$ is the steady state solution of the ode given a trainable parameter $x$ and a condition $c$. The trainable parameter $x$ and the "external" condition $c$ are set using parameterlens and conditionsetter lenses which act on ode.p respectively.

SteadyStateObjective also provides Jacobian in the fg!(F, G, x) form required by Optim.only_fg!.

source

SteadyStateFit.optimize — Function.

optimize(sso::SteadyStateObjective, [x0, method, options])

Optimize a SteadyStateObjective sso using Optim. The rest of the positional and keyword arguments are passed to Optim.optimize.

source