Home

SteadyStateFit.jl

SteadyStateFit.SteadyStateObjective

DiffEqBase.solve

SteadyStateFit.SteadyStateObjective — Type.

SteadyStateObjective(
    loss,
    ode::ODEProblem,
    conditions::AbstractVector,
    conditionsetter::Lens,
    parameterlens::Lens,
    steadystatesolver = NLSolver();
    nlsolvecallback = _ -> nothing,
)

SteadyStateObjective defines an objective function that x ↦ F(x) that computes

loss(states :: AbstractVector{TU}, conditions :: AbstractVector{TC}, sso) :: Real

where

states is a vector of steady state u(x, c :: TC) :: TU of the ode computed given a trainable parameter x and each c in conditions,

conditions is the argument passed to SteadyStateObjective, and

sso is a SteadyStateObjective such that sso.p contains the parameter x for which the objective is evaluated; i.e. sso.p is set(ode.p, parameterlens, x).

The trainable parameter x and the "external" condition c are set using parameterlens and conditionsetter lenses which act on ode.p respectively.

SteadyStateObjective also provides Jacobian in the fg!(F, G, x) form required by Optim.only_fg!.

source

DiffEqBase.solve — Function.

solve(sso::SteadyStateObjective, [method, options, x0]) :: SteadyStateFitResult

Optimize a SteadyStateObjective sso using Optim. The rest of the positional and keyword arguments are passed to Optim.optimize.

source